requirements.txt
azure-identity==1.6.0
mysql-connector-python==8.0.28
requests==2.26.0